refactor: restructure repo for multi-tool GH Pages layout
Move CLI merge tool into cli-merge/ subdirectory so the repo can host
multiple tools. Add root index.html landing page. Update test paths to
tests/cli-merge/ and update deno.json + CI workflow to glob tests/.

Structure:
  index.html            ← landing page
  cli-merge/
    index.html          ← merge tool
    src/{parser,output}.js
  tests/cli-merge/      ← tests mirroring tool layout
