+++ .github/workflows/test.yml
@@ -17,7 +17,7 @@ jobs:
           deno-version: v2.x
 
       - name: Check formatting
-        run: deno fmt --check src/ tests/
+        run: deno fmt --check tests/
 
       - name: Run tests
-        run: deno test --allow-all --no-check tests/parser.test.ts tests/output.test.ts
+        run: deno test --allow-all --no-check tests/
