+++ cli-merge/src/output.js
