+++ cli-merge/src/parser.js
